bump geom and change back to float partially

+++ libfive.cpp
@@ -18,46 +18,46 @@ using namespace libfive;
 
 extern "C" {
 
-double libfive_tree_eval_d_spread(libfive_tree tree, double x, double y, double z) {
-  libfive_vec3 p = {(float)x, (float)y, (float)z};
+double libfive_tree_eval_d_spread(libfive_tree tree, float x, float y, float z) {
+  libfive_vec3 p = {x, y, z};
   return libfive_tree_eval_f(tree, p);
 }
 
-bool libfive_tree_save_mesh_spread(libfive_tree tree, double lx, double ux, double ly, double uy, double lz, double uz,
-                                   double res, const char* f);
+bool libfive_tree_save_mesh_spread(libfive_tree tree, float lx, float ux, float ly, float uy, float lz, float uz,
+                                   float res, const char* f);
  
-libfive_contours* libfive_tree_render_slice_spread(libfive_tree tree, double lx, double ux, double ly, double uy, 
-                                                   double z, double res);
+libfive_contours* libfive_tree_render_slice_spread(libfive_tree tree, float lx, float ux, float ly, float uy, 
+                                                   float z, float res);
 
-libfive_pixels* libfive_tree_render_pixels_spread(libfive_tree tree, double lx, double ux, double ly, double uy, 
-                                                  double z, double res);
+libfive_pixels* libfive_tree_render_pixels_spread(libfive_tree tree, float lx, float ux, float ly, float uy, 
+                                                  float z, float res);
 
-bool libfive_tree_save_slice_spread(libfive_tree tree, double lx, double ux, double ly, double uy, double z,
-                                    double res, const char* f);
+bool libfive_tree_save_slice_spread(libfive_tree tree, float lx, float ux, float ly, float uy, float z,
+                                    float res, const char* f);
 
-libfive_mesh* libfive_tree_render_mesh_spread(libfive_tree tree, double lx, double ux, double ly, double uy, double lz, double uz,
-                                    double res);
+libfive_mesh* libfive_tree_render_mesh_spread(libfive_tree tree, float lx, float ux, float ly, float uy, float lz, float uz,
+                                              float res);
   
-bool libfive_tree_save_mesh_spread(libfive_tree tree, double lx, double ux, double ly, double uy, double lz, double uz,
-                                   double res, const char* f) {
-  libfive_region3 region = {{(float)lx, (float)ux}, {(float)ly, (float)uy}, {(float)lz, (float)uz}};
+bool libfive_tree_save_mesh_spread(libfive_tree tree, float lx, float ux, float ly, float uy, float lz, float uz,
+                                   float res, const char* f) {
+  libfive_region3 region = {{lx, ux}, {ly, uy}, {lz, uz}};
   return libfive_tree_save_mesh(tree, region, res, f);
 }
  
-libfive_contours* libfive_tree_render_slice_spread(libfive_tree tree, double lx, double ux, double ly, double uy, 
-                                                  double z, double res) {
-  libfive_region2 region = {{(float)lx, (float)ux}, {(float)ly, (float)uy}};
+libfive_contours* libfive_tree_render_slice_spread(libfive_tree tree, float lx, float ux, float ly, float uy, 
+                                                   float z, float res) {
+  libfive_region2 region = {{lx, ux}, {ly, uy}};
   return libfive_tree_render_slice(tree, region, z, res);
 }
 
-libfive_pixels* libfive_tree_render_pixels_spread(libfive_tree tree, double lx, double ux, double ly, double uy, 
-                                                  double z, double res) {
-  libfive_region2 region = {{(float)lx, (float)ux}, {(float)ly, (float)uy}};
+libfive_pixels* libfive_tree_render_pixels_spread(libfive_tree tree, float lx, float ux, float ly, float uy, 
+                                                  float z, float res) {
+  libfive_region2 region = {{lx, ux}, {ly, uy}};
   return libfive_tree_render_pixels(tree, region, z, res);
 }
 
-bool libfive_tree_save_slice_spread(libfive_tree tree_data, double lx, double ux, double ly, double uy, double z,
-                                 double res, const char* f) {
+bool libfive_tree_save_slice_spread(libfive_tree tree_data, float lx, float ux, float ly, float uy, float z,
+                                    float res, const char* f) {
   Region<2> region({lx, ly}, {ux, uy}, Region<2>::Perp(z));
   BRepSettings settings;
   settings.min_feature = 1/res;
@@ -68,9 +68,9 @@ bool libfive_tree_save_slice_spread(libfive_tree tree_data, double lx, double ux
   return 0;
 }
 
-libfive_mesh* libfive_tree_render_mesh_spread(libfive_tree tree, double lx, double ux, double ly, double uy, double lz, double uz,
-                                    double res) {
-  libfive_region3 region = {{(float)lx, (float)ux}, {(float)ly, (float)uy}, {(float)lz, (float)uz}};
+libfive_mesh* libfive_tree_render_mesh_spread(libfive_tree tree, float lx, float ux, float ly, float uy, float lz, float uz,
+                                              float res) {
+  libfive_region3 region = {{lx, ux}, {ly, uy}, {lz, uz}};
   return libfive_tree_render_mesh(tree, region, res);
 }
